move MyCancerGenome implementation into react-mutation-mapper (fix cBioPortal/cbioportal/issues/6991)
