+++ packages/config/rollup.config.js
@@ -2,6 +2,7 @@ import typescript from 'rollup-plugin-typescript2';
 import commonjs from '@rollup/plugin-commonjs';
 // import external from 'rollup-plugin-peer-deps-external';
 import autoExternal from 'rollup-plugin-auto-external';
+import json from '@rollup/plugin-json';
 import postcss from 'rollup-plugin-postcss';
 import resolve from '@rollup/plugin-node-resolve';
 import sourcemaps from 'rollup-plugin-sourcemaps';
@@ -49,6 +50,7 @@ export default function getRollupOptions(
                 rollupCommonJSResolveHack: true,
                 clean: true,
             }),
+            json(),
             commonjs(),
             resolve(),
             sourcemaps(),
