+++ packages/react-mutation-mapper/src/store/DefaultMutationMapperStore.ts
@@ -25,6 +25,7 @@ import { ApplyFilterFn, FilterApplier } from '../model/FilterApplier';
 import { Gene } from '../model/Gene';
 import { Mutation } from '../model/Mutation';
 import MutationMapperStore from '../model/MutationMapperStore';
+import { IMyCancerGenomeData } from '../model/MyCancerGenome';
 import { PfamDomain, PfamDomainRange } from '../model/Pfam';
 import { PostTranslationalModification } from '../model/PostTranslationalModification';
 import {
@@ -45,6 +46,7 @@ import {
     groupMutationsByProteinStartPos,
     uniqueGenomicLocations,
 } from '../util/MutationUtils';
+import { getMyCancerGenomeData } from '../util/MyCancerGenomeUtils';
 import {
     defaultOncoKbIndicatorFilter,
     groupOncoKbIndicatorDataByMutations,
@@ -161,6 +163,8 @@ class DefaultMutationMapperStore implements MutationMapperStore {
             this.config.groupFilters
         );
     }
+
+    public readonly myCancerGenomeData: IMyCancerGenomeData = getMyCancerGenomeData();
 
     @computed
     public get mutations(): Mutation[] {
